SentimentApp/requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-form-utils==1.0.3
pytz==2018.7
selenium==3.141.0
urllib3==1.24.1
